FSM Refactor
Refactored project to use the State Pattern.
Each state is implemented as a subclass of the State class.
Each event is represented by a method that is overriden by each state for the relevant events.
Methods for events that trigger a state transition return a pointer to that state which is then used as the current state.