Added alarm functionality
Added additional states for setting the alarm time and for when the alarm goes off.
A couple of small alarm handler functions were added to be called from the RTC alarm ISR to add an event to the queue.
The alarm edit state is entered by holding button 2 from the time display state.
The alarm state is entered from the time display state when the alarm event occurs.
The alarm state can be left by pressing any button.