Added Buzzer
Added a passive buzzer that can be driven with a square wave and a class to implment playing a tone.
Added functions to play a tune on the buzzer to act as the alarm sound.
Alarm sound is activated when the alarm state is entered and stopped when the alarm state is left.